@@ -8,14 +8,14 @@
 	<properties>
 		<quarkus.platform.artifact-id>quarkus-bom</quarkus.platform.artifact-id>
 		<quarkus.platform.group-id>io.quarkus</quarkus.platform.group-id>
-		<quarkus.platform.version>2.11.2.Final</quarkus.platform.version>
-		<compiler-plugin.version>3.8.1</compiler-plugin.version>
-		<surefire-plugin.version>3.0.0-M7</surefire-plugin.version>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.target>11</maven.compiler.target>
-		<maven.compiler.parameters>true</maven.compiler.parameters>
-	</properties>
+                <quarkus.platform.version>3.9.5</quarkus.platform.version>
+                <compiler-plugin.version>3.11.0</compiler-plugin.version>
+                <surefire-plugin.version>3.2.5</surefire-plugin.version>
+                <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+                <maven.compiler.source>21</maven.compiler.source>
+                <maven.compiler.target>21</maven.compiler.target>
+                <maven.compiler.parameters>true</maven.compiler.parameters>
+        </properties>
 
 	<dependencyManagement>
 		<dependencies>
