@@ -4,10 +4,10 @@ import java.sql.Timestamp;
 import java.util.Calendar;
 import java.util.List;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.transaction.Transactional;
-import javax.ws.rs.WebApplicationException;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.transaction.Transactional;
+import jakarta.ws.rs.WebApplicationException;
 
 import org.eclipse.microprofile.jwt.JsonWebToken;
 
