@@ -1,8 +1,8 @@
 package com.opyruso.propertiesmanager.services.impl;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.ws.rs.WebApplicationException;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.WebApplicationException;
 
 import org.apache.http.HttpStatus;
 import org.eclipse.microprofile.config.ConfigProvider;
