@@ -2,16 +2,16 @@ package com.opyruso.propertiesmanager.data.entity.pk;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 
 @Embeddable
 public class ApplicationPK implements Serializable {
 
 	private static final long serialVersionUID = -6857143440338021902L;
 
-	@Column(name = "app_id", nullable = false)
-	private String appId;
+       @Column(name = "app_id", nullable = false, length = 100)
+       private String appId;
 
 	@Override
 	public int hashCode() {
