@@ -2,16 +2,16 @@ package com.opyruso.propertiesmanager.data.entity.pk;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 
 @Embeddable
 public class GlobalVariablePK implements Serializable {
 
 	private static final long serialVersionUID = -6607564208223563517L;
 
-	@Column(name = "globalvariable_key", nullable = false)
-	private String globalVariableKey;
+       @Column(name = "globalvariable_key", nullable = false, length = 100)
+       private String globalVariableKey;
 
 	@Override
 	public int hashCode() {
