@@ -2,22 +2,22 @@ package com.opyruso.propertiesmanager.data.entity.pk;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 
 @Embeddable
 public class PropertiesFilePK implements Serializable {
 
 	private static final long serialVersionUID = 3884974871349693735L;
 
-	@Column(name = "app_id", nullable = false)
-	private String appId;
+       @Column(name = "app_id", nullable = false, length = 100)
+       private String appId;
 
-	@Column(name = "num_version", nullable = false)
-	private String numVersion;
+       @Column(name = "num_version", nullable = false, length = 100)
+       private String numVersion;
 
-	@Column(name = "filename", nullable = false)
-	private String filename;
+       @Column(name = "filename", nullable = false, length = 100)
+       private String filename;
 
 	@Override
 	public int hashCode() {
