@@ -3,7 +3,7 @@ package com.opyruso.propertiesmanager.constants;
 import java.util.HashMap;
 import java.util.Map;
 
-import javax.inject.Singleton;
+import jakarta.inject.Singleton;
 
 import org.eclipse.microprofile.config.ConfigProvider;
 
