@@ -1,6 +1,6 @@
 package com.opyruso.propertiesmanager.services;
 
-import javax.ws.rs.WebApplicationException;
+import jakarta.ws.rs.WebApplicationException;
 
 import com.opyruso.propertiesmanager.data.entity.UserRightsDemand;
 
