@@ -3,8 +3,8 @@ package com.opyruso.propertiesmanager.utils;
 import java.util.HashMap;
 import java.util.Map;
 
-import javax.inject.Inject;
-import javax.inject.Singleton;
+import jakarta.inject.Inject;
+import jakarta.inject.Singleton;
 
 import com.opyruso.propertiesmanager.api.entity.ApiEnvironment;
 import com.opyruso.propertiesmanager.constants.EnvironmentConfig;
